feat: added events page and basic styling to Event & Events components
