+++ app/events/page.tsx
@@ -1,0 +1,6 @@
+import Events from './events'
+export default function Page() {
+    return (
+        <Events />
+    )
+}
